pyproject.toml
[tool.poetry]
name = "hubspot-s3-uploader"
version = "0.2.0"
authors = ["Jing Su <sujing1894@gmail.com>"]
description = "Sync ingredients from S3 JSON files to HubSpot, orchestrated via Google Sheet."
readme = "README.md"
packages = [{ include = "hubspot_s3_uploader" }]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34"
requests = "^2.32"
python-dotenv = "^1.0"
gspread = "^6.0"
google-auth = "^2.29"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"